# Stream symbols during compaction.