@bwplotka 2019-12-17
Anything interesting on the resource-side of things? I see this includes https://github.com/prometheus/prometheus/pull/6461 so curious of e2e results (: