@brian-brazil 2019-12-17
The symbol table in the benchmark is only a few kB, so not visible at all.