@brian-brazil 2019-12-17
> If we are happy with 3x code complexity for 10% gain, I think it's good to go, but I am worried about maintainability of this for mere mortals (:

Any speed benefits are an unexpected bonus, the RAM savings are the aim here.