@johncming 2019-12-24
use `TestIndexRW_Postings` for test, the value of symbol table is `8` bytes. 

@brian-brazil why `8` bytes here?

https://github.com/prometheus/prometheus/blob/2b653ee2301c2658f8ffec0160c0dc2ad92a9441/tsdb/index/index.go#L1304-L1306