@brian-brazil 2019-12-24
A uint64 is 8 bytes.